sets group view
