+++ apps/templates/pages/create_group.html
@@ -24,7 +24,7 @@
     <div class="card">
       <div class="card-body create-group-card-body">
         <p class="create-group-box-msg">Create a New Group</p>
-        <form method="POST" action="{{ url_for('authentication_blueprint.create_group') }}">
+        <form method="POST" action="{{ url_for('home_blueprint.create_group') }}">
           {{ form.hidden_tag() }}
           <p class="create-group-box-msg">
             {% if msg %}
@@ -78,7 +78,7 @@
               <button type="submit" class="btn btn-primary btn-block" name="create">Create Group</button>
             </div>
             <div class="col-8 text-right">
-              <a href="{{ url_for('home_blueprint.view_group') }}" class="btn btn-secondary">Cancel</a>
+              <a href="{{ url_for('home_blueprint.view_groups') }}" class="btn btn-secondary">Cancel</a>
             </div>
           </div>
         </form>
