+++ apps/templates/pages/edit_group.html
@@ -106,7 +106,7 @@
 
               <div class="row">
                 <button class="btn btn-primary" type="submit">Save</button>
-                <a href="{{ url_for('home_blueprint.view_group') }}" class="btn btn-secondary mx-2" role="button" aria-pressed="true">Cancel</a>
+                <a href="{{ url_for('home_blueprint.view_groups') }}" class="btn btn-secondary mx-2" role="button" aria-pressed="true">Cancel</a>
               </div>
               <!-- /.row -->
             </form>
